fixed mistakes

+++ spring-10/src/main/java/ru/otus/hw/services/BookServiceImpl.java
@@ -34,7 +34,8 @@ public class BookServiceImpl implements BookService {
     @Override
     @Transactional(readOnly = true)
     public List<BookDto> findAll() {
-        return bookRepository.findAll().stream()
+        return bookRepository.findAll()
+                .stream()
                 .map(bookConverter::fromDomainObject)
                 .toList();
     }
@@ -54,6 +55,9 @@ public class BookServiceImpl implements BookService {
     @Override
     @Transactional
     public void deleteById(long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Book with id %d not found".formatted(id));
+        }
         bookRepository.deleteById(id);
     }
